Check type before receiveType.  Fixes #429.
When the user provides the DragSource/DropTarget type as a string or
array, the type is checked immediately.  When the user provides the
type as a function, it's not checked until it's added to the registry.
If this function returns null (and the type has not yet been set), then
the type will never be checked, resulting in cryptic error messages.